# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ApunBazar E-Commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  originalPrice?: number | null;
  categoryId: number;
  categoryName: string;
  categorySlug: string;
  imageUrl: string;
  images: string[];
  stock: number;
  rating: number;
  reviewCount: number;
  featured: boolean;
  artisan?: string | null;
  origin?: string | null;
  tags: string[];
  createdAt: string;
}

export interface ProductsListResponse {
  products: Product[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface CreateProductBody {
  name: string;
  description: string;
  price: number;
  originalPrice?: number | null;
  categoryId: number;
  imageUrl: string;
  images?: string[];
  stock: number;
  featured?: boolean;
  artisan?: string | null;
  origin?: string | null;
  tags?: string[];
}

export interface UpdateProductBody {
  name?: string;
  description?: string;
  price?: number;
  originalPrice?: number | null;
  categoryId?: number;
  imageUrl?: string;
  images?: string[];
  stock?: number;
  featured?: boolean;
  artisan?: string | null;
  origin?: string | null;
  tags?: string[];
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  description?: string | null;
  imageUrl?: string | null;
  productCount: number;
}

export interface CreateCategoryBody {
  name: string;
  slug: string;
  description?: string | null;
  imageUrl?: string | null;
}

export interface CartItem {
  id: number;
  productId: number;
  product: Product;
  quantity: number;
  sessionId: string;
}

export interface Cart {
  items: CartItem[];
  total: number;
  itemCount: number;
}

export interface AddToCartBody {
  sessionId: string;
  productId: number;
  quantity: number;
}

export interface UpdateCartItemBody {
  quantity: number;
}

export interface WishlistItem {
  id: number;
  productId: number;
  product: Product;
  sessionId: string;
  addedAt: string;
}

export interface AddToWishlistBody {
  sessionId: string;
  productId: number;
}

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  productImage: string;
  price: number;
  quantity: number;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending: "pending",
  confirmed: "confirmed",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export type OrderPaymentMethod =
  (typeof OrderPaymentMethod)[keyof typeof OrderPaymentMethod];

export const OrderPaymentMethod = {
  cod: "cod",
  online: "online",
} as const;

export type OrderPaymentStatus =
  (typeof OrderPaymentStatus)[keyof typeof OrderPaymentStatus];

export const OrderPaymentStatus = {
  pending: "pending",
  paid: "paid",
  failed: "failed",
} as const;

export interface Order {
  id: number;
  orderNumber: string;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  shippingAddress: string;
  shippingCity: string;
  shippingState: string;
  shippingPincode: string;
  items: OrderItem[];
  subtotal: number;
  shippingFee: number;
  total: number;
  status: OrderStatus;
  paymentMethod: OrderPaymentMethod;
  paymentStatus: OrderPaymentStatus;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface OrdersListResponse {
  orders: Order[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export type CreateOrderBodyItemsItem = {
  productId: number;
  quantity: number;
  price: number;
};

export type CreateOrderBodyPaymentMethod =
  (typeof CreateOrderBodyPaymentMethod)[keyof typeof CreateOrderBodyPaymentMethod];

export const CreateOrderBodyPaymentMethod = {
  cod: "cod",
  online: "online",
} as const;

export interface CreateOrderBody {
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  shippingAddress: string;
  shippingCity: string;
  shippingState: string;
  shippingPincode: string;
  items: CreateOrderBodyItemsItem[];
  paymentMethod: CreateOrderBodyPaymentMethod;
  notes?: string | null;
  sessionId?: string;
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  pending: "pending",
  confirmed: "confirmed",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export type UpdateOrderStatusBodyPaymentStatus =
  (typeof UpdateOrderStatusBodyPaymentStatus)[keyof typeof UpdateOrderStatusBodyPaymentStatus];

export const UpdateOrderStatusBodyPaymentStatus = {
  pending: "pending",
  paid: "paid",
  failed: "failed",
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
  paymentStatus?: UpdateOrderStatusBodyPaymentStatus;
}

export interface AdminStats {
  totalRevenue: number;
  totalOrders: number;
  totalProducts: number;
  totalCustomers: number;
  pendingOrders: number;
  revenueThisMonth: number;
  ordersThisMonth: number;
  averageOrderValue: number;
}

export interface TopProduct {
  id: number;
  name: string;
  imageUrl: string;
  totalSold: number;
  revenue: number;
  categoryName: string;
}

export interface CategorySales {
  categoryName: string;
  totalOrders: number;
  revenue: number;
  percentage: number;
}

export type ListProductsParams = {
  category?: string;
  search?: string;
  minPrice?: number;
  maxPrice?: number;
  sort?: ListProductsSort;
  page?: number;
  limit?: number;
  featured?: boolean;
};

export type ListProductsSort =
  (typeof ListProductsSort)[keyof typeof ListProductsSort];

export const ListProductsSort = {
  price_asc: "price_asc",
  price_desc: "price_desc",
  newest: "newest",
  popular: "popular",
} as const;

export type GetCartParams = {
  sessionId: string;
};

export type GetWishlistParams = {
  sessionId: string;
};

export type RemoveFromWishlistParams = {
  sessionId: string;
};

export type ListOrdersParams = {
  email?: string;
  status?: string;
  page?: number;
  limit?: number;
};

export type GetRecentOrdersParams = {
  limit?: number;
};

export type GetTopProductsParams = {
  limit?: number;
};