# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ApunBazar E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get cart by session
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cart contents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/{itemId}:
    put:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
  /admin/recent-orders:
    get:
      operationId: getRecentOrders
      tags: [admin]
      summary: Recent orders for admin dashboard
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /admin/top-products:
    get:
      operationId: getTopProducts
      tags: [admin]
      summary: Top selling products
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
          nullable: true
        categoryId:
          type: integer
        categoryName:
          type: string
        categorySlug:
          type: string
        imageUrl:
          type: string
        images:
          type: array
          items:
            type: string
        stock:
          type: integer
        rating:
          type: number
        reviewCount:
          type: integer
        featured:
          type: boolean
        artisan:
          type: string
          nullable: true
        origin:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required: [id, name, description, price, categoryId, categoryName, categorySlug, imageUrl, images, stock, rating, reviewCount, featured, tags, createdAt]
...
    CartItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        product:
          $ref: "#/components/schemas/Product"
        quantity:
          type: integer
        sessionId:
          type: string
      required: [id, productId, product, quantity, sessionId]
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: number
        itemCount:
          type: integer
      required: [items, total, itemCount]
    AddToCartBody:
      type: object
      properties:
        sessionId:
          type: string
        productId:
          type: integer
        quantity:
          type: integer
      required: [sessionId, productId, quantity]
    UpdateCartItemBody:
      type: object
      properties:
        quantity:
          type: integer
      required: [quantity]
...
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImage:
          type: string
        price:
          type: number
        quantity:
          type: integer
      required: [id, productId, productName, productImage, price, quantity]
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        shippingAddress:
          type: string
        shippingCity:
          type: string
        shippingState:
          type: string
        shippingPincode:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        shippingFee:
          type: number
        total:
          type: number
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
        paymentMethod:
          type: string
          enum: [cod, online]
        paymentStatus:
          type: string
          enum: [pending, paid, failed]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, orderNumber, customerName, customerEmail, customerPhone, shippingAddress, shippingCity, shippingState, shippingPincode, items, subtotal, shippingFee, total, status, paymentMethod, paymentStatus, createdAt, updatedAt]
...
    TopProduct:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        imageUrl:
          type: string
        totalSold:
          type: integer
        revenue:
          type: number
        categoryName:
          type: string
      required: [id, name, imageUrl, totalSold, revenue, categoryName]
...